+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+- Implemented pagination in `SearchOffersAsync` method.
+
 ## [0.1.1-preview] - 2025-08-29
 
 - Updated definition of `OfferFilter` class.